change type of creationDate add creationDate in UiState rename HabitSave

+++ app/src/main/java/com/example/newapppp/data/remote/habit/HabitDone.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class HabitDone(
     @SerialName(value = "date")
-    val creationDate: Long,
+    val creationDate: Int,
     @SerialName(value = "habit_uid")
     val id: String
 )
